Adjust legacy map editor overview formatting
