+++ src/apps/legacy/map-editor/index.ts
@@ -1,0 +1,2 @@
+// src/apps/legacy/map-editor/index.ts
+export * from "./view";
